requirements.txt
Django==5.2.1
requests==2.32.3
pywebview==5.1
